Add Encoders class
This class is able to count the encoder clicks from the Pololu hall
effect encoders.  It counts on each rise and fall of both quadature
signals from each sensor.  This results in 12 counts per rotation of
the motor which should be 120 counts per rotation of my wheel because
of the 10:1 gearbox.