Add MPU-9150 documentation
This is the 9DoF IMU board that I am using from Sparkfun.  Currently I
am just taking advantage of the gyro and accelerometer sensors and
ignoring the magnetometer.  I could have just used the MPU-6050 device
in this project but I got these more capable ones to have as a porting
target for the Kalman filter that I wrote last year.