Instantiate g_enocders later
It must be instantiated after other InterruptIn object so that its
constructor will find the InterruptIn handler already in the
interrupt vector to remember before replacing.  When the encoder
handler gets called it will call the remembered interrupt handler if
there are still any GPIO interrupts to be handled after the encoder
ones are processed.