Can now detect radio issues
I have added a hasTimedOut() method to PwmIn so that it can be called
to detect if the mbed is no longer receiving pulses from the radio
receiver.  This can be used later to put the robot in emergency
shutdown mode if the radio is turned off or connection is otherwise
lost.