Added brake on radio timeout option
I now set the BRAKE_ON_RADIO_TIMEOUT macro to brake the motors when the
radio connection has been lost.  Previously the code would try to
maintain a 0 velocity when the radio was turned off.  This might be Ok
in general for the real bot but for my initial testing it would put
more stress on the motor when I just wanted to give it a break from a
failing scenario.