Add pitch calibration mode
This mode will oscillate the platform back and forth.  The motor PWM
output and the IMU pitch samples taken every 0.01 seconds will be
logged to a 32k RAM buffer for ~40 seconds and the saved to a
PITCH.CSV file in the LocalFileSystem.  This commit contains an
example of such a PITCH.CSV file created during my initial test run.

The PITCH_CALIBRATE_ANGLE and PITCH_CALIBRATE_PWM values can be
set to determine what motor power and platform pitch angle should
be used for the oscillation.