Initial commit of mriprog utility
This utility takes advantage of the MRI debug monitor running on a
LPC1768 device to program new firmware into the device.  If the
UART has been bridged to WiFi using a ESP8266 chip then this
programming can be done remotely.  I use the
https://github.com/jeelabs/esp-link firmware on a ESP8266 board to
provide this bridging.

Example command line:
  mriprog 192.168.0.129:23 ../esp8266test/LPC1768/esp8266.bin
This uploads the esp8266test image to the device which has MRI
talking over WiFi on TCP/IP address 192.168.0.129, port 23.