Fix PID implementation
I was initializing m_Kd incorrectly in PID::setTuningParameters().  It
should be Kc * Td and not Kc / Td.

It also doesn't make sense for PID::compute() to adjust the integral to
eliminate integral windup when Ti == 0.0f.  In that scenario, the
windup advoidance code was the only thing which would update the
integral and there was nothing else to correct it over time.  This
would lead to a bad offset being applied when testing with a Kp
parameter only.