mriprog: Default to only erasing updated FLASH sectors
Previously mriprog would erase all of FLASH before writing in the new
contents. Now it defaults to only erasing the sectors just before it
writes to each one. If it doesn't write to the sector then it doesn't
erase it. This is good if there is another bootloader in FLASH that
you don't want to erase while uploading your newest firmware with
mriprog.

Added a new "-e" command line flag which enables the old behaviour of
erasing all FLASH sectors before starting to write in the new contents.

It is best to use the -e flag when possible since it is almost twice as
fast. I actually had to increase the serial timeout from 1 second to
2 seconds to account for the time it can take to erase and write to
16k of FLASH when erasing sectors as it progresses.