Stop motors when entering debugger
I have added implementations of the __mriPlatform_EnteringDebuggerHook
and __mriPlatform_LeavingDebuggerHook functions to be called by the MRI
debug monitor when it takes control and then later releases the CPU.

The motors are set to a braking condition when entering the debugger if
the g_motorsSetup global has been set indicating that the global
constructors have been run and the motor/PID/Ticker objects
initialized. If the setup hasn't completed yet then just enter debug
control without doing anything special. The ticker which runs the PID
code is also stopped when entering debug control and then re-enabled
upon exit. If this isn't done then the Ticker object will starve the
CPU while it tries to catch up for all of time missed while under
debugger control.