Fix PwmIn time out detection code
This commit fixes a bug where the code would previously detect that no
PWM pulse had been detected for awhile (time out) when the radio was
actually still connected and sending pulses as it should. This bug was
caused by the fact that PwmIn previously used a mbed Timer object to
track the time since the last edge had been detected. The methods on
this class do not support concurrency so sometimes it would underflow
during subtraction and return a very large elapsed time:
* My code would call Timer::read_us() to determine the time since the
  last edge was detected to determine if this elapsed time exceeded the
  specified m_usecTimeOut.
* Timer::read_us() calls Timer::slicetime(). This method includes this
  code:
    int Timer::slicetime() {
        if (_running) {
            return ticker_read(_ticker_data) - _start;
        } else {
* That slicetime() method contains the code which could underflow in
  this scenario:
  * Calls ticker_read() to read current time.
  * Radio generates an edge that causes an interrupt to fire:
    * The interrupt handler updates _start to a new (and more recent
      value.)
    * Interrupt handler returns to allow slicetime() to continue
      running.
  * Timer::slicetime() continues and reads in new value of _start
    member which is now newer than the value previously read via
    the ticker_read() call. The subtraction then generates a very
    large unsigned value (would be negative if signed). The large
    value looks like a timeout.

I updated the PwmIn class to no longer use the mbed Timer object and
instead I just record the 32-bit time returned from us_ticker_read()
each time an edge is detected. I call the elapsedTimeSinceLastEdge()
whenever I want to determine how much time has elapsed since the last
edge. This method detects if the m_lastEdgeTime changes during the
calculation and retries when it does.

I will remove the printf() debugging code included in this commit in a
future commit. It was this printf() of the large elapsed time which
clued me in on the cause of the random timeouts.