MRI UART has highest priority
Previously findMriUart() assumed that the UART used by MRI would be
running at an interrupt priority of 0, the highest. There are some
scenarios where it might be ok to run the MRI UART at a level lower
than 0 but even then it should still be the UART running at the highest
priority.