Could call realloc() on non-malloc'ed memory
Need to zero out the new ElfRegions as they are allocated to make sure
that the ELF_REGION_FLAGS_ALLOCATED flag isn't set. Sometimes I would
get a crash during elf file parsing when the new CRT detected the
invalid realloc() call. This might be the cause of the problems that
I had seen before since it could have left the heap in a weird state
before if the CRT wasn't catching it like it is now.