Send CTRL+C to stop running process
Before a running process would be stopped by the first '+' character
sent and MRI would respond with a T packet when mriprog expected the
qSupported response. Now the CTRL+C will generate the T packet, then
the '+' will ACK that T packet, and finally the qSupported command and
response can be processed successfully.